requirements.txt
streamlit
pillow
tensorflow-cpu
numpy
scikit-learn
matplotlib
seaborn